LibArchive is playing along finally
Have a basic test application working which prints the contents of
archive.zip to the console. lso created a separate debug dll and library
pair for debug versions of libarchive, since it was so persuasive. Only
the debug version of the libarchive library seems to work. ( Tried doing
a release build of the test application, never again)