Fixed GetStructure()
Places a 'children array' into fileStructure class variable. The array
an array for all the folders in the archive, and stores the name of the
folder in fileStructure[i][0], and any indexes (of fileStructure array -
as strings!) of children the folder might have in subsequent cells, eg.
fileStructure[i][1] == "1" means fileStructure[1] is the parent of i.