Extended Python Script
Script can now take 'open' or 'write' along with a file name of a zip,
and, in the case of 'write' extra file names to write to the zip.
Currently can not create empty folders in the zip. Open will just open
and read the contents of the zip, but will not display the results.